handlePress, handleLongPress refactored
