+++ App.js
@@ -1,11 +1,11 @@
 import store from './store';
-import GestureWrapper from './Layout/GestureWrapper';
+import LayoutWrapper from './Layout/LayoutWrapper';
 import {Provider} from 'react-redux';
 
 export default function App() {
   return (
     <Provider store={store}>
-      <GestureWrapper />
+      <LayoutWrapper />
     </Provider>
   );
 }
